+++ CHANGELOG.md
@@ -1,15 +1,28 @@
+## 0.1.4
+
+Renamed functions for registering factories with parameters 
+and deprecated the old ones. Below you can find the full map of 
+changes:
+
+| Deprecated | New |
+| ---- | ----- |
+| `registerFactoryWithOneParameter` | `registerFactoryWithParam` | 
+| `registerFactoryWithTwoParameters` | `registerFactoryWithTwoParams` |
+| `registerFactoryAsyncWithOneParameter` | `registerFactoryAsyncWithParam` |
+| `registerFactoryAsyncWithTwoParameters` | `registerFactoryAsyncWithTwoParams` |
+
 ## 0.1.3
 
-- Fix documentation regarding lazy initialization.
+Fix documentation regarding lazy initialization.
 
 ## 0.1.2
 
-- Add factories with params
+Add factories with params
 
 ## 0.1.1
 
-- Add async factories
+Add async factories
 
 ## 0.1.0
 
-- Initial version of Toor 🚀
+Initial version of Toor 🚀